[Feat, Refactor] 좋아요 수 조회 기능 추가, 리팩토링 #52

+++ src/main/java/com/example/fantreehouse/domain/feedlike/repository/FeedLikeRepository.java
@@ -12,5 +12,6 @@ public interface FeedLikeRepository extends JpaRepository<FeedLike, Long>{
 
     List<FeedLike> findAllFeedLikeByFeedId(Long artistFeedId);
 
-    int countByFeedId(Long id);
+    Long countByFeedId(Long id);
+
 }
